+++ financial-connections/src/main/java/com/stripe/android/financialconnections/FinancialConnectionsSheetResultCallback.kt
@@ -1,0 +1,8 @@
+package com.stripe.android.financialconnections
+
+/**
+ * Callback that is invoked when a [FinancialConnectionsSheetResult] is available.
+ */
+fun interface FinancialConnectionsSheetResultCallback {
+    fun onFinancialConnectionsSheetResult(financialConnectionsSheetResult: FinancialConnectionsSheetResult)
+}