+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 ## x.x.x - xxxx-xx-xx
 
+## 20.1.0 - 2022-04-18
+This release includes several Payments and PaymentSheet bug fixes.
+
 ### Payments (`com.stripe:stripe-android`)
 * [ADDED] [4874](https://github.com/stripe/stripe-android/pull/4874) `us_bank_account` PaymentMethod is now available for ACH Direct Debit payments, including APIs to collect customer bank information (requires Connections SDK) and verify microdeposits.
 * [FIXED] [4875](https://github.com/stripe/stripe-android/pull/4875) fix postal code callback not firing when enabled
@@ -29,7 +32,7 @@ changes the public API for CardImageVerificationSheet and releases Identity SDK.
 ### Identity (`com.stripe:identity`)
 * [ADDED] [4820](https://github.com/stripe/stripe-android/pull/4820) Release Stripe's Identity SDK.
 
-### Connections (`com.stripe:connection`)
+### Financial Connections (`com.stripe:financial-connections`)
 * [ADDED] [4818](https://github.com/stripe/stripe-mandroid/pull/4818) Connections SDK can be optionally included to support ACH Direct Debit payments.
 
 ### CardScan (`com.stripe:stripecardscan`)